pyproject.toml
[project]
name = "lserve-anima"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-keyvault-secrets>=4.9.0",
    "joblib>=1.5.1",
    "litserve>=0.2.10",
    "numpy>=2.2.6",
    "torch>=2.7.0",
]
